backend/requirements.txt
fastapi
uvicorn
yt-dlp
supabase
python-dotenv
requests
pydantic
httpx
